+++ saos-persistence/src/main/java/pl/edu/icm/saos/persistence/repository/JudgmentRepository.java
@@ -9,13 +9,13 @@ import pl.edu.icm.saos.persistence.model.Judgment;
 /**
  * @author Łukasz Dumiszewski
  */
-public interface JudgmentRepository extends JudgmentCommonRepository<Judgment>, JpaRepository<Judgment, Integer>, JudgmentRepositoryCustom {
+public interface JudgmentRepository extends JudgmentCommonRepository<Judgment>, JpaRepository<Judgment, Long>, JudgmentRepositoryCustom {
 
     
     /**
      * Deletes fast all the {@link Judgment}s with the given ids using just a few jpql/sql queries
      */
-    public void delete(List<Integer> judgmentIds);
+    public void delete(List<Long> judgmentIds);
     
     
 }
