+++ saos-persistence/src/main/java/pl/edu/icm/saos/persistence/repository/LawJournalEntryRepository.java
@@ -8,7 +8,7 @@ import pl.edu.icm.saos.persistence.model.LawJournalEntry;
  * @author Łukasz Dumiszewski
  */
 
-public interface LawJournalEntryRepository extends JpaRepository<LawJournalEntry, Integer> {
+public interface LawJournalEntryRepository extends JpaRepository<LawJournalEntry, Long> {
 
     
     public LawJournalEntry findOneByYearAndJournalNoAndEntry(int year, int journalNo, int entry);
